Rename SQLite helper module to avoid package collision
